Board: Creates a board with max tiles of 100. Made a Board method that creates the board and tiles used in BoardGameApp.
BoardGame: Added getboard and getdice for BoardGameApp. moved the dice roll to the Die class.

BoardGameApp: Creates a playable application of the board game, in console. had problems making a playturn() method so its manually rolls and moves the player

Changed dice to properly create dice, and roll the dice.

changed die to properly roll each die for the player.

changed player so it can actually move in the board game and method for checking what tile the player is currently on. also creates a dice for each player.

created id's for each tile, getter method for the boardgame and a setter method for the board class.

Most of this code is made in a small hurry, as i wanted to get things done, so proper code hygiene and public/private statements on methods/field are not final.